pyproject.toml
[project]
name = "ai-casino"
version = "0.1.0"
description = "Agentic stock trading system with multi-agent architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Technical Analysis
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "numpy>=1.26.0",

    # Sentiment Analysis
    "transformers>=4.36.0",
    "torch>=2.1.0",

    # LLM & Agents
    "litellm>=1.50.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # Market Data
    "yfinance>=0.2.40",
    "alpha-vantage>=2.3.1",
    "requests>=2.31.0",

    # Trading
    "alpaca-py>=0.29.0",
    "backtesting>=0.3.3",

    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "loguru>=0.7.2",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
